moved cuda_stream to root
